Application_Module/Emergency_webserver/requirements.txt
Flask==1.1.2      
Flask-Cors==3.0.9 
paho-mqtt==1.5.1
psycopg2==2.8.6
